requirements.txt
flask
pdf2image
opencv-python-headless
numpy
img2pdf
pyzbar
Pillow
matplotlib
google-genai
